// src/types/Post.ts
export type Post = {
  ageRange?: null;
  bookmarkCnt?: number;
  commentCnt?: number;
  connectStatus?: 'wait';
  content?: string;
  createdAt?: number;
  gender?: null;
  id?: number;
  isBlocked?: 'N';
  isDeleted?: 'N';
  locationCode?: number;
  memberId?: number;
  nickname?: string;
  prayedCnt?: number;
  profileImage?: string;
  region1depthName?: string;
  region2depthName?: string;
  region3depthName?: string;
  reportedCnt?: number;
  title?: string;
  updatedAt?: number;
  views?: number;
};